+++ lib/24_03_13/abstract/intangible_asset.dart
@@ -1,18 +1,18 @@
 import 'asset.dart';
 
-abstract class TangibleAsset extends Asset {
+abstract class IntangibleAsset extends Asset {
   //타입의 경우 추후 enum같은 것을 활용한 네임스페이스를 사용할것 아직안배움..
   final String _type;
   final int _amount;
 
-  TangibleAsset(
+  IntangibleAsset(
       {required super.name,
       required super.price,
       required String type,
       required int amount})
       : _type = type,
         _amount = amount,
-        super(isTangible: true);
+        super(isTangible: false);
 
   String get type => _type;
